refactor: moved input-style-generator test-cases to constants file
